Changed expected manifest id for insecure-cookie and avoid-persistent-cookies samples
