+++ scala/src/detectors/avoid-persistent-cookies/compliant.scala
@@ -3,20 +3,18 @@
 *  SPDX-License-Identifier: Apache-2.0
 */
 
-import play.api.mvc.{Cookie => PlayCookie, _}
-import javax.servlet.http.{Cookie => ServletCookie, HttpServletResponse}
+import javax.servlet.http.Cookie
+import javax.servlet.http.HttpServletResponse
 
-class AvoidPersistentCookiesCompliant @Inject()(cc: ControllerComponents) extends AbstractController(cc) {
-
-  // {fact rule=avoid-persistent-cookies@v1.0 defects=0}
-  def compliant(res: HttpServletResponse, name: String, value: String, secure: Boolean = true, maxAge: Int = 60, httpOnly: Boolean = true): Unit = {
-    val cookie = new Cookie("key", "value")
-    // Compliant: Cookie `setSecure` method is set to true.
-    cookie.setSecure(true)
-    cookie.setMaxAge(60)
-    cookie.setHttpOnly(true)
-    res.addCookie(cookie)
-  }
-  // {/fact}
-
+class AvoidPersistentCookiesCompliant {
+    // {fact rule=avoid-persistent-cookies@v1.0 defects=0}
+    def compliant(res: HttpServletResponse): Unit = {
+        val cookie = new Cookie("key", "value")
+        cookie.setSecure(true)
+        cookie.setHttpOnly(true)
+        // Compliant: MaxAge set to one week.
+        cookie.setMaxAge(604800)
+        res.addCookie(cookie)
+    }
+    // {/fact}
 }
