+++ scala/src/detectors/insecure-cookie/compliant.scala
@@ -3,18 +3,20 @@
 *  SPDX-License-Identifier: Apache-2.0
 */
 
-import javax.servlet.http.Cookie
-import javax.servlet.http.HttpServletResponse
+import play.api.mvc.{Cookie => PlayCookie, _}
+import javax.servlet.http.{Cookie => ServletCookie, HttpServletResponse}
 
-class InsecureCookieCompliant {
-    // {fact rule=insecure-cookie@v1.0 defects=0}
-    def compliant(res: HttpServletResponse): Unit = {
-        val cookie = new Cookie("key", "value")
-        cookie.setSecure(true)
-        cookie.setHttpOnly(true)
-        // Compliant: MaxAge set to one week.
-        cookie.setMaxAge(604800)
-        res.addCookie(cookie)
-    }
-    // {/fact}
+class InsecureCookieCompliant @Inject()(cc: ControllerComponents) extends AbstractController(cc) {
+
+  // {fact rule=insecure-cookie@v1.0 defects=0}
+  def compliant(res: HttpServletResponse, name: String, value: String, secure: Boolean = true, maxAge: Int = 60, httpOnly: Boolean = true): Unit = {
+    val cookie = new Cookie("key", "value")
+    // Compliant: Cookie `setSecure` method is set to true.
+    cookie.setSecure(true)
+    cookie.setMaxAge(60)
+    cookie.setHttpOnly(true)
+    res.addCookie(cookie)
+  }
+  // {/fact}
+
 }
