+++ scala/src/detectors/insecure-cookie/non-compliant.scala
@@ -3,18 +3,20 @@
 *  SPDX-License-Identifier: Apache-2.0
 */
 
-import javax.servlet.http.Cookie
-import javax.servlet.http.HttpServletResponse
+import play.api.mvc.{Cookie => PlayCookie, _}
+import javax.servlet.http.{Cookie => ServletCookie, HttpServletResponse}
 
-class InsecureCookieNoncompliant {
-    // {fact rule=insecure-cookie@v1.0 defects=1}
-    def nonCompliant(res: HttpServletResponse): Unit = {
-        val cookie = new Cookie("key", "value")
-        cookie.setSecure(true)
-        cookie.setHttpOnly(true)
-        // Noncompliant: MaxAge set to one year.
-        cookie.setMaxAge(31536000)
-        res.addCookie(cookie)
-    }
-    // {/fact}
+class InsecureCookieNoncompliant @Inject()(cc: ControllerComponents) extends AbstractController(cc) {
+
+  // {fact rule=insecure-cookie@v1.0 defects=1}
+  def nonCompliant(res: HttpServletResponse, name: String, value: String, secure: Boolean = true, maxAge: Int = 60, httpOnly: Boolean = true): Unit = {
+    val cookie = new Cookie("key", "value")
+    // Noncompliant: Cookie `setSecure` method is set to false.
+    cookie.setSecure(false)
+    cookie.setMaxAge(60)
+    cookie.setHttpOnly(true)
+    res.addCookie(cookie)
+  }
+  // {/fact}
+
 }
